refactor: Move ui Python modules to gradience/ui/ directory
