+++ gradience/modules/preset.py
@@ -19,7 +19,7 @@
 import json
 import os
 
-from ..settings_schema import settings_schema
+from gradience.settings_schema import settings_schema
 from .utils import buglog, to_slug_case
