+++ gradience/ui/error.py
@@ -18,7 +18,7 @@
 
 from gi.repository import Gtk
 
-from .constants import rootdir
+from gradience.constants import rootdir
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/error.ui")
