+++ gradience/ui/no_plugin_pref.py
@@ -1,6 +1,6 @@
 from gi.repository import Gtk, Adw
 
-from .constants import rootdir
+from gradience.constants import rootdir
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/no_plugin_pref.ui")
