+++ gradience/ui/preferences.py
@@ -19,16 +19,16 @@
 
 from gi.repository import Gtk, Adw
 
-from .constants import rootdir
-from .modules.flatpak_overrides import (
+from gradience.constants import rootdir
+from gradience.modules.flatpak_overrides import (
     create_gtk_user_override,
     remove_gtk_user_override,
 )
-from .modules.flatpak_overrides import (
+from gradience.modules.flatpak_overrides import (
     create_gtk_global_override,
     remove_gtk_global_override,
 )
-from .modules.utils import buglog
+from gradience.modules.utils import buglog
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/preferences.ui")
