+++ gradience/ui/welcome.py
@@ -1,27 +1,45 @@
+# welcome.py
+#
+# Change the look of Adwaita, with ease
+# Copyright (C) 2022  Gradience Team
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+
 import sys
 import time
 
 from gi.repository import Gtk, Adw, Gio
 
-from .modules.run_async import RunAsync
-from .modules.utils import buglog
-from .modules.flatpak_overrides import create_gtk_user_override
-from .constants import rootdir, app_id, rel_ver
+from gradience.modules.run_async import RunAsync
+from gradience.modules.utils import buglog
+from gradience.modules.flatpak_overrides import create_gtk_user_override
+from gradience.constants import rootdir, app_id, rel_ver
 
 
-@Gtk.Template(resource_path=f"{rootdir}/ui/share_window.ui")
-class GradienceShareWindow(Adw.Window):
-    __gtype_name__ = "GradienceShareWindow"
+@Gtk.Template(resource_path=f"{rootdir}/ui/welcome.ui")
+class GradienceWelcomeWindow(Adw.Window):
+    __gtype_name__ = "GradienceWelcomeWindow"
 
     settings = Gtk.Settings.get_default()
+
+    carousel = Gtk.Template.Child()
 
     btn_close = Gtk.Template.Child()
     btn_back = Gtk.Template.Child()
     btn_next = Gtk.Template.Child()
     btn_install = Gtk.Template.Child()
     btn_agree = Gtk.Template.Child()
-
-    carousel = Gtk.Template.Child()
 
     switch_system = Gtk.Template.Child()
     switch_adw_gtk3 = Gtk.Template.Child()
@@ -37,19 +55,23 @@ class GradienceShareWindow(Adw.Window):
 
     carousel_pages = [
         "welcome",  # 0
-        "gradience",  # 1
-        "configure",  # 2
-        "download",  # 3
-        "finish",  # 4
+        "release",  # 1
+        "agreement",  # 2
+        "gradience",  # 3
+        "configure",  # 4
+        "download",  # 5
+        "finish",  # 6
     ]
 
     page_welcome = Gtk.Template.Child()
     page_release = Gtk.Template.Child()
 
-    def __init__(self, window, **kwargs) -> None:
+    def __init__(self, window, update=False, **kwargs) -> None:
         super().__init__(**kwargs)
 
         self.set_transient_for(window)
+
+        self.update = update
 
         # common variables and references
         self.window = window
@@ -62,11 +84,16 @@ class GradienceShareWindow(Adw.Window):
         self.btn_back.connect("clicked", self.previous_page)
         self.btn_next.connect("clicked", self.next_page)
         self.btn_install.connect("clicked", self.install_runner)
-
+        self.btn_agree.connect("clicked", self.agree)
         self.settings.connect(
             "notify::gtk-application-prefer-dark-theme", self.theme_changed
         )
         self.connect("close-request", self.quit)
+
+        if self.update:
+            self.page_welcome.set_title(_("Thanks for updating Gradience!"))
+
+        self.page_release.set_title(f"Gradience {rel_ver}")
 
         self.btn_close.set_sensitive(False)
 
@@ -92,9 +119,15 @@ class GradienceShareWindow(Adw.Window):
         """
         page = self.get_page(index)
 
+        self.carousel.set_interactive(True)
         if page == "finish":
             self.btn_back.set_visible(False)
             self.btn_next.set_visible(False)
+            self.carousel.set_interactive(False)
+        elif page == "agreement":
+            self.btn_back.set_visible(True)
+            self.btn_next.set_visible(False)
+            self.btn_agree.set_visible(True)
             self.carousel.set_interactive(False)
         elif page == "download":
             self.btn_back.set_visible(True)
@@ -104,15 +137,43 @@ class GradienceShareWindow(Adw.Window):
         elif page == "welcome":
             self.btn_back.set_visible(False)
             self.btn_next.set_visible(True)
-            self.carousel.set_interactive(True)
         else:
             self.btn_back.set_visible(True)
             self.btn_next.set_visible(True)
             self.btn_install.set_visible(False)
             self.carousel.set_interactive(True)
 
+    def agree(self, widget):
+        self.window.last_opened_version = self.window.settings.set_string(
+            "last-opened-version", rel_ver
+        )
+
+        if self.update:
+            self.btn_close.set_sensitive(True)
+            self.label_skip.set_visible(False)
+            self.next_page(index=5)
+        else:
+            self.next_page()
+
     def quit(self, *args):
         self.destroy()
+        sys.exit()
+
+    # TODO: Add adw-gtk3 check
+    def check_adw_gtk3(self, *args):
+        buglog("check if adw-gtk3 installed")
+        return True
+
+    def adw_gtk3(self):
+        if not self.check_adw_gtk3():  # install
+            buglog("install adw-gtk3")
+
+    def configure_system(self):
+        buglog("configure system")
+        self.allow_flatpak_theming_user_toggled()
+
+    def allow_flatpak_theming_user_toggled(self, *args):
+        create_gtk_user_override(self, self.gio_settings, "gtk4")
 
     def install_runner(self, widget):
         def set_completed(result, error=False):
@@ -125,7 +186,11 @@ class GradienceShareWindow(Adw.Window):
         self.set_deletable(False)
 
         def install():
-            print("Installing Gradience…")
+            if self.switch_adw_gtk3.get_active():
+                self.adw_gtk3()
+
+            if self.switch_system.get_active():
+                self.configure_system()
 
         RunAsync(self.pulse)
         RunAsync(
